+++ CHANGELOG.md
@@ -57,6 +57,8 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 - feat: optimize `RequestTimeoutException`, when a timeout occurs, if the server returns some information, you can read them.
 - refactor!: mark `Asset#create(String, String, String)` as deprecated, use `Asset.createNonNativeAsset(String, String)` or `Asset.createNativeAsset()` instead.
 - refactor!: remove `LiquidityPoolID`. Use `String` to represent the liquidity pool ID.
+- refactor!: `TransactionPreconditions#TransactionPreconditions(LedgerBounds, Long, BigInteger, long, List, TimeBounds)` has been removed, use `TransactionPreconditions#TransactionPreconditions(TimeBounds, LedgerBounds, Long, BigInteger, long, List)` instead.
+- refactor: Set the default value of `TransactionPreconditions.extraSigners` to `new ArrayList<>()`, it is not nullable.
 
 ## 0.44.0
 ### Update